Update ui.R
On the Intro Tab:Changed the titles, further polishing. 
On the Upload tab: Only changed title of the tab. 
On the Explore tab: changed order of the tables, gave the text a different focus, and polished the text and titles
On the Help the birds tab: Polished the text, added a link for the current grassroots conservation projects EDGE has.